Refactor controllers to have one route handler per controller
Rewrote controllers so that only the module method route is exported
from a given controller.  This method in turn calls methods defined
in the controller, which is a object literal containing methods to
handle different operations for the controller urls.  Added a
signupController which includes sanitization of inputs to prevent
script injection.