Improve handling of session cache information
Abstract out creation of session hash.  Go to mongo if session
cache not present when deserializing user.  Set expiration value
for session cache when creating or accessing session cache.  On
logout remove the session cache from redis.  Note this could be
further improved by saving the session id in the user collection
in Mongo -- if a user can have more than one open session, then
the sessions would be stored as an array in the document.