Add delete idea and user methods to support testing.
In IdeaController, added deleteIdea and deleteIdeas api calls and routes.  Fix issue with getIdea and editIdea that permitted a user to see or modify another user's idea.  Added return value of the idea id for addIdea.  In UserController,  added deleteUser method and route.  Fixed issues with editUser so user can modify any part of the profile.